[refactor]: split patient intake app service by intake, triage and summary concerns
